requirements.txt
Django==4.2.10
Pillow==10.2.0
psycopg2-binary==2.9.9
django-cleanup==8.0.0
django-imagekit==4.1.0
django-meta==2.5.0
django-sitemaps==1.0.0
django-debug-toolbar==4.2.0
whitenoise==6.6.0
django-compressor==4.4
django-allauth==0.58.2
python-dotenv==1.0.1
django-widget-tweaks==1.5.0
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
django-storages==1.14.2
django-cors-headers==4.3.1
django-filter==23.5
django-rest-framework==0.1.0
django-celery-beat==2.5.0
celery==5.3.6
redis==5.0.1
gunicorn==21.2.0
django-environ==0.11.2
django-taggit==5.0.1
django-ckeditor==6.7.0
django-redis==5.4.0
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-celery-results==2.5.1
sentry-sdk==1.39.1
django-stubs==4.2.7
django-stubs-ext==4.2.7
mypy==1.8.0
black==24.1.1
flake8==7.0.0
isort==5.13.2
coverage==7.4.1
pytest==8.0.0
pytest-django==4.7.0
factory-boy==3.3.0
faker==22.6.0
django-admin-sortable2==2.2.8